CSN/requirements.txt
networkx==2.4
scikit-learn==0.21.1
scipy==1.4.1
pytorch==1.11
torchsummary==1.5.1